// product/frontend/DNAustria/src/app/api/model/locationReplyDto.ts
/**
 * DNAustria.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressReplyDto } from './addressReplyDto';


export interface LocationReplyDto { 
    id?: number;
    name: string;
    latitude?: number;
    longitude?: number;
    address: AddressReplyDto | null;
}